+++ docs/search-core.md
@@ -17,8 +17,10 @@
 |  --- | --- |
 |  [AppliedQueryFilterType](./search-core.appliedqueryfiltertype.md) | Represents the type of [AppliedQueryFilter](./search-core.appliedqueryfilter.md) applied to a search. |
 |  [BuiltInFieldType](./search-core.builtinfieldtype.md) | Possible built-in field types for [DirectAnswer.fieldType](./search-core.directanswer.fieldtype.md)<!-- -->. |
+|  [CloudRegion](./search-core.cloudregion.md) | Defines the cloud region of the API domains. |
 |  [DirectAnswerType](./search-core.directanswertype.md) | Represents the type of direct answer. |
 |  [Direction](./search-core.direction.md) | The direction of a sort. |
+|  [Environment](./search-core.environment.md) | Defines the environment of the API domains. |
 |  [ErrorType](./search-core.errortype.md) | Identifier for the type of error causing the failure. |
 |  [FilterCombinator](./search-core.filtercombinator.md) | Indicates how child filters in a [StaticFilter](./search-core.staticfilter.md) should be combined. |
 |  [LocationBiasMethod](./search-core.locationbiasmethod.md) | The method used to determine the location. |
