+++ src/index.ts
@@ -1,6 +1,6 @@
 // main symbols
 export { SearchCore } from './SearchCore';
 export { provideCore } from './provideCore';
-export { provideEndpoints, SandboxEndpoints } from './constants';
+export { provideEndpoints, Environment, CloudRegion, SandboxEndpoints } from './provideEndpoints';
 export * from './models';
 export * from './services';
