+++ tests/infra/QuestionSubmissionServiceImpl.ts
@@ -3,7 +3,7 @@ import { HttpServiceMock } from '../mocks/HttpServiceMock';
 import { HttpService } from '../../src/services/HttpService';
 import { SearchConfigWithDefaulting } from '../../src/models/core/SearchConfig';
 import { ApiResponseValidator } from '../../src/validation/ApiResponseValidator';
-import { defaultApiVersion, provideEndpoints } from '../../src/constants';
+import { defaultApiVersion, provideEndpoints } from '../../src/provideEndpoints';
 
 const baseCoreConfig = {
   apiKey: 'anApiKey',
